refactors stream code; moves PointBuffer to its own file; adds time related utilities
